🚧 Work-in-progress rewrite of the project
Completely changing the API to be less perscriptive of how the consumer may manage their sources.

Rewriting the theming and layout to sepearte this from the render layer. This won't be nessessary for anything that is exporting to a format which already specifies the style (like HTML snippets).

Most of this rewrite will be moved to a seperate library so that I can focus of supporting DocFX and .SVG generation can become a secondary feature.